+++ src/Client.php
@@ -11,9 +11,9 @@ use SmartAssert\ServiceClient\Exception\InvalidModelDataException;
 use SmartAssert\ServiceClient\Exception\InvalidResponseContentException;
 use SmartAssert\ServiceClient\Exception\InvalidResponseDataException;
 use SmartAssert\ServiceClient\Exception\NonSuccessResponseException;
-use SmartAssert\ServiceClient\Payload\UrlEncodedPayload;
 use SmartAssert\SourcesClient\Model\SourceInterface;
 use SmartAssert\SourcesClient\RequestHandler\FileRequestHandler;
+use SmartAssert\SourcesClient\RequestHandler\SourceMutationHandler;
 
 class Client
 {
@@ -23,6 +23,7 @@ class Client
         private readonly SourceFactory $sourceFactory,
         private readonly ExceptionFactory $exceptionFactory,
         private readonly FileRequestHandler $fileRequestHandler,
+        private readonly SourceMutationHandler $sourceMutationHandler,
     ) {
     }
 
@@ -36,7 +37,7 @@ class Client
      */
     public function createFileSource(string $token, string $label): SourceInterface
     {
-        return $this->makeFileSourceMutationRequest($token, $label, null);
+        return $this->sourceMutationHandler->createFileSource($token, $label);
     }
 
     /**
@@ -50,7 +51,7 @@ class Client
      */
     public function updateFileSource(string $token, string $sourceId, string $label): SourceInterface
     {
-        return $this->makeFileSourceMutationRequest($token, $label, $sourceId);
+        return $this->sourceMutationHandler->updateFileSource($token, $sourceId, $label);
     }
 
     /**
@@ -71,7 +72,7 @@ class Client
         string $path,
         ?string $credentials,
     ): SourceInterface {
-        return $this->makeGitSourceMutationRequest($token, $label, $hostUrl, $path, $credentials, null);
+        return $this->sourceMutationHandler->createGitSource($token, $label, $hostUrl, $path, $credentials);
     }
 
     /**
@@ -94,7 +95,7 @@ class Client
         string $path,
         ?string $credentials,
     ): SourceInterface {
-        return $this->makeGitSourceMutationRequest($token, $label, $hostUrl, $path, $credentials, $sourceId);
+        return $this->sourceMutationHandler->updateGitSource($token, $sourceId, $label, $hostUrl, $path, $credentials);
     }
 
     /**
@@ -199,88 +200,6 @@ class Client
     {
         $response = $this->serviceClient->sendRequestForJsonEncodedData(
             $this->requestFactory->createSourceRequest('DELETE', $token, $sourceId)
-        );
-
-        if (!$response->isSuccessful()) {
-            throw $this->exceptionFactory->createFromResponse($response);
-        }
-
-        $source = $this->sourceFactory->create($response->getData());
-        if (null === $source) {
-            throw InvalidModelDataException::fromJsonResponse(SourceInterface::class, $response);
-        }
-
-        return $source;
-    }
-
-    /**
-     * @param non-empty-string      $token
-     * @param non-empty-string      $label
-     * @param null|non-empty-string $sourceId
-     *
-     * @throws ClientExceptionInterface
-     * @throws HttpResponseExceptionInterface
-     * @throws InvalidModelDataException
-     */
-    private function makeFileSourceMutationRequest(
-        string $token,
-        string $label,
-        ?string $sourceId,
-    ): SourceInterface {
-        return $this->makeSourceMutationRequest($token, ['type' => 'file', 'label' => $label], $sourceId);
-    }
-
-    /**
-     * @param non-empty-string      $label
-     * @param non-empty-string      $token
-     * @param non-empty-string      $hostUrl
-     * @param non-empty-string      $path
-     * @param null|non-empty-string $credentials
-     * @param null|non-empty-string $sourceId
-     *
-     * @throws ClientExceptionInterface
-     * @throws HttpResponseExceptionInterface
-     * @throws InvalidModelDataException
-     */
-    private function makeGitSourceMutationRequest(
-        string $token,
-        string $label,
-        string $hostUrl,
-        string $path,
-        ?string $credentials,
-        ?string $sourceId,
-    ): SourceInterface {
-        $payload = [
-            'type' => 'git',
-            'label' => $label,
-            'host-url' => $hostUrl,
-            'path' => $path,
-        ];
-
-        if (is_string($credentials)) {
-            $payload['credentials'] = $credentials;
-        }
-
-        return $this->makeSourceMutationRequest($token, $payload, $sourceId);
-    }
-
-    /**
-     * @param non-empty-string      $token
-     * @param null|non-empty-string $sourceId
-     * @param array<mixed>          $payload
-     *
-     * @throws ClientExceptionInterface
-     * @throws HttpResponseExceptionInterface
-     * @throws InvalidModelDataException
-     */
-    private function makeSourceMutationRequest(
-        string $token,
-        array $payload,
-        ?string $sourceId,
-    ): SourceInterface {
-        $response = $this->serviceClient->sendRequestForJsonEncodedData(
-            $this->requestFactory->createSourceRequest(is_string($sourceId) ? 'PUT' : 'POST', $token, $sourceId)
-                ->withPayload(new UrlEncodedPayload($payload))
         );
 
         if (!$response->isSuccessful()) {
